Make test proper.
